+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+* Simple WebSocket connection
+* Basic UI display
+
 ## [Alpha0.1] - 2026-07-10
 
 ### Added